+++ src/GTGT/wrappers.py
@@ -2,7 +2,6 @@ from .models import TranscriptModel, BedModel
 from .provider import Provider
 from .ensembl import lookup_transcript as lookup_transcript_ens
 from .ucsc import lookup_knownGene
-from . import Bed
 
 
 def lookup_transcript(provider: Provider, transcript_id: str) -> TranscriptModel: